Run update-po preparing for unifying configure.ac
Update all *.po files using "make update-po" to prepare for the
change to a single configure.ac script.

That single configure.ac commit will be easier to read if the
corresponding *.po changes just deal with the single configure.ac
change, and not with everything unrelated which just happens to
have happened earlier.